+++ inc/classes/class-robotstxt.php
@@ -9,101 +9,151 @@
  * @link       /LICENSE
  */
 
-namespace MsRobotstxtManager;
+namespace RobotstxtManager;
 
 if ( false === defined( 'ABSPATH' ) ) {
 	exit;
 }
 
 /**
- * Display Robots.txt File If Called
+ * Robots.txt
  */
 final class Robotstxt {
 	/**
-	 * Check File Being Called
+	 * Maybe start robots.txt display
 	 */
 	public function __construct() {
-		if ( false !== strpos( filter_input( INPUT_SERVER, 'REQUEST_URI', FILTER_SANITIZE_URL ), 'robots.txt' ) ) {
-			$this->robotstxt();
+		if ( true === $this->is_robotstxt_file() ) {
+			$this->display();
 		}
 	}
 
 
 	/**
-	 * Display Robots.txt File
+	 * Display robots.txt file
 	 */
-	private function robotstxt() {
-		/*
-		 * If Multisite is enabled.
-		 * https://developer.wordpress.org/reference/functions/is_multisite/
-		 */
-		if ( true === is_multisite() ) {
-			/*
-			 * Retrieve the current site ID.
-			 * https://developer.wordpress.org/reference/functions/get_current_blog_id/
-			 */
-			$site_id = get_current_blog_id();
+	private function display() {
+		$robotstxt   = $this->get_robotstxt();
+		$blog_public = \get_option( 'blog_public' );
 
-			/*
-			 * Switch the current blog.
-			 * https://developer.wordpress.org/reference/functions/switch_to_blog/
-			 */
-			switch_to_blog( $site_id );
+		if ( '0' === $blog_public ) {
+			$robotstxt = "Disallow: /\n";
 		}
 
-		/*
-		 * Retrieves an option value based on an option name.
-		 * https://developer.wordpress.org/reference/functions/get_option/
-		 */
-		$website_option = get_option( MS_ROBOTSTXT_MANAGER_PLUGIN_NAME );
-		$robotstxt_file = '';
-
-		// Ignore If Disabled.
-		if ( true !== empty( $website_option['disable'] ) ) {
-			return;
-		}
-
-		// Robots.txt Set.
-		if ( true !== empty( $website_option['robotstxt'] ) && true === empty( $robotstxt_file ) ) {
-			$robotstxt_file = $website_option['robotstxt'];
-		}
-
-		// Override Robots.txt Set, Use Append As Robots.txt File.
-		if ( true !== empty( $website_option['override'] ) ) {
-			$robotstxt_file = $website_option['append'];
-		}
-
-		// Display Robots.txt File.
-		if ( true !== empty( $robotstxt_file ) ) {
+		if ( true !== empty( $robotstxt ) ) {
 			header( 'Status: 200 OK', true, 200 );
-			header( 'Content-type: text/plain; charset=' . get_bloginfo( 'charset' ) );
+			header( 'Content-Type: text/plain; charset=utf-8' );
 
 			/**
-			 * Fires when displaying the robots.txt file.
+			 * Escaping for HTML blocks.
 			 *
-			 * @since 2.1.0
+			 * @source https://developer.wordpress.org/reference/functions/esc_html/
 			 */
-			do_action( 'do_robotstxt' );
-
-			// Get Status.
-			$public = get_option( 'blog_public' );
-
-			if ( '0' === $public ) {
-				$output .= "Disallow: /\n";
-			} else {
-				$output = $robotstxt_file;
-			}
-
-			/**
-			 * Filters the robots.txt output.
-			 *
-			 * @since 3.0.0
-			 *
-			 * @param string $output Robots.txt output.
-			 * @param bool   $public Whether the site is considered "public".
-			 */
-			echo apply_filters( 'robots_txt', $output, $public );
+			echo \esc_html( $robotstxt );
 			exit;
 		}
 	}
-}//end class
+
+
+	/**
+	 * Get robots.txt file
+	 *
+	 * @return string
+	 */
+	private function get_robotstxt() : string {
+		/**
+		 * Retrieve the current site ID.
+		 *
+		 * @source https://developer.wordpress.org/reference/functions/get_current_blog_id/
+		 */
+		$blog_id = \get_current_blog_id();
+
+		if ( $blog_id > 0 &&
+			true === function_exists( 'switch_to_blog' ) ) {
+			/**
+			 * Switch the current blog.
+			 *
+			 * @source https://developer.wordpress.org/reference/functions/switch_to_blog/
+			 */
+			\switch_to_blog( $blog_id );
+		}
+
+		/**
+		 * Retrieves an option value based on an option name.
+		 *
+		 * @source https://developer.wordpress.org/reference/functions/get_option/
+		 */
+		$settings = \get_option( ROBOTSTXT_MANAGER_PLUGIN_NAME );
+
+		if ( $blog_id > 0 &&
+			true === function_exists( 'restore_current_blog' ) ) {
+			/**
+			 * Restore the current blog, after calling restore_current_blog().
+			 *
+			 * @source https://developer.wordpress.org/reference/functions/restore_current_blog/
+			 */
+			\restore_current_blog();
+		}
+
+		$robotstxt = '';
+
+		if ( true !== empty( $settings['disable'] ) ) {
+			return null;
+		}
+
+		if ( true !== empty( $settings['robotstxt'] ) ) {
+			$robotstxt = $settings['robotstxt'];
+		}
+
+		if ( true !== empty( $robotstxt ) ) {
+			$public = \get_option( 'blog_public' );
+
+			if ( '0' === $public ) {
+				$robotstxt = "Disallow: /\n";
+			}
+		}
+
+		if ( true !== empty( $user_robotstxt ) ) {
+			$robotstxt = $user_robotstxt;
+		}
+
+		return (string) $robotstxt;
+	}
+
+
+	/**
+	 * Check if called file is robots.txt file
+	 */
+	private function is_robotstxt_file() : bool {
+		if ( true === filter_has_var( INPUT_SERVER, 'REQUEST_URI' ) ) {
+			$filename = filter_input(
+				INPUT_SERVER,
+				'REQUEST_URI',
+				FILTER_UNSAFE_RAW,
+				FILTER_NULL_ON_FAILURE
+			);
+		} else {
+			if ( isset( $_SERVER['REQUEST_URI'] ) ) {
+				/**
+				 * Remove slashes from a string or array of strings.
+				 *
+				 * @source https://developer.wordpress.org/reference/functions/wp_unslash/
+				 */
+				$request_uri = \wp_unslash( $_SERVER['REQUEST_URI'] );
+				$filename    = filter_var(
+					$request_uri,
+					FILTER_UNSAFE_RAW,
+					FILTER_NULL_ON_FAILURE
+				);
+			} else {
+				$filename = null;
+			}
+		}
+
+		if ( '/robots.txt' === $filename || 'robotst.txt' === $filename ) {
+			return true;
+		}
+
+		return false;
+	}
+}